New Player art added - Done.
Max said that he'll probably improve this art again during the polish stage but he wanted to get this added in now rather than obsess over it forever.
I also noticed that the background stars are moving a too fast and in the wrong direction so that's another thing to deal with!