Mirror custom collider & level 1 basic design - Done?
Level 1 is still called Sandbox and is using the placeholder art but it should be fine for now.
I (Charlie) have planned out a few more levels but I will wait for the new mirror art first. Well Sunday will probably be the last day that I wait.
Also level 1 is quite hard and only beatable with rotation and careful aiming. I think I need a simpler level 1 and then this can become level 2.
I spent quite a bit of today fixing collision errors with the walls and creating invisible walls to use to constrain the Player's movement.
With these constraints it's possible to stop the Player easily winning.