New mirror art added in - Done.
Max made some new art for the mirrors.
It's been added in with new colliders and actually works better with the physics than the old mirror. Cool stuff! :)
There was also a bug with the mirror collider but that's fixed now.
I'm(Charlie) going to commit these changes while everything works and then I'll go and work on level design.